Create Minify.py
As assigned, this is my attempt to write a function Minifying the XML File ! 

Since spaces and newlines (\n) are actually characters that can 
increase the size of an XML document. This feature should aim at decreasing the size of an XML file (compressing it) by deleting the whitespaces and indentations.